refactor: 일부 구조 수정
