+++ src/pages/battle/short.tsx
@@ -5,7 +5,7 @@ import { useState } from 'react';
 
 import AuthRequiredPage from '@/components/auth/AuthRequiredPage';
 import Battle from '@/components/battle/detail/Battle';
-import { useGetBattle } from '@/components/battle/detail/useGetBattle';
+import { useGetBattle } from '@/components/battle/useGetBattle';
 import BottomNav from '@/components/common/BottomNav';
 import Genres from '@/components/common/Genres';
 import Header from '@/components/common/Header';
